+++ src/main/java/com/getpcpanel/ui/SoundDeviceImportFactory.java
@@ -1,6 +1,6 @@
 package com.getpcpanel.ui;
 
-import com.getpcpanel.util.SoundDevice;
+import com.getpcpanel.cpp.AudioDevice;
 
 import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
@@ -11,29 +11,29 @@ import javafx.scene.input.MouseEvent;
 import javafx.scene.input.TransferMode;
 import javafx.util.Callback;
 
-public class SoundDeviceImportFactory implements Callback<ListView<SoundDevice>, ListCell<SoundDevice>> {
+public class SoundDeviceImportFactory implements Callback<ListView<AudioDevice>, ListCell<AudioDevice>> {
     public static final DataFormat JAVA_FORMAT = new DataFormat("application/x-java-serialized-object");
 
     private static final String DROP_HINT_STYLE_ABOVE = "-fx-border-color: #eea82f; -fx-border-width: 2 0 0 0; -fx-padding: 1 7 3 7;";
 
     private static final String DROP_HINT_STYLE_BELLOW = "-fx-border-color: #eea82f; -fx-border-width: 0 0 2 0; -fx-padding: 3 7 1 7";
 
-    private ListCell<SoundDevice> dropZone;
+    private ListCell<AudioDevice> dropZone;
 
-    private SoundDevice draggedItem;
+    private AudioDevice draggedItem;
 
-    private final ListView<SoundDevice> listView;
+    private final ListView<AudioDevice> listView;
 
-    public SoundDeviceImportFactory(ListView<SoundDevice> listView) {
+    public SoundDeviceImportFactory(ListView<AudioDevice> listView) {
         this.listView = listView;
         setupListView(listView);
     }
 
     @Override
-    public ListCell<SoundDevice> call(ListView<SoundDevice> listView) {
-        var cell = new ListCell<SoundDevice>() {
+    public ListCell<AudioDevice> call(ListView<AudioDevice> listView) {
+        var cell = new ListCell<AudioDevice>() {
             @Override
-            protected void updateItem(SoundDevice item, boolean empty) {
+            protected void updateItem(AudioDevice item, boolean empty) {
                 super.updateItem(item, empty);
                 if (item == null) {
                     setText("");
@@ -50,7 +50,7 @@ public class SoundDeviceImportFactory implements Callback<ListView<SoundDevice>,
         return cell;
     }
 
-    private void dragDetected(MouseEvent event, ListCell<SoundDevice> treeCell, ListView<SoundDevice> treeView) {
+    private void dragDetected(MouseEvent event, ListCell<AudioDevice> treeCell, ListView<AudioDevice> treeView) {
         draggedItem = treeCell.getItem();
         var db = treeCell.startDragAndDrop(TransferMode.MOVE);
         var content = new ClipboardContent();
@@ -60,11 +60,11 @@ public class SoundDeviceImportFactory implements Callback<ListView<SoundDevice>,
         event.consume();
     }
 
-    private void dragOver(DragEvent event, ListCell<SoundDevice> treeCell, ListView<SoundDevice> treeView) {
+    private void dragOver(DragEvent event, ListCell<AudioDevice> treeCell, ListView<AudioDevice> treeView) {
         if (!event.getDragboard().hasContent(JAVA_FORMAT))
             return;
         var thisItem = treeCell.getItem();
-        var dropContent = (SoundDevice) event.getDragboard().getContent(JAVA_FORMAT);
+        var dropContent = (AudioDevice) event.getDragboard().getContent(JAVA_FORMAT);
         if (event.getGestureSource().getClass().getEnclosingClass().getName().contains("SoundDeviceImportFactory")) {
             if (thisItem == draggedItem)
                 return;
@@ -83,7 +83,7 @@ public class SoundDeviceImportFactory implements Callback<ListView<SoundDevice>,
         }
     }
 
-    private void drop(DragEvent event, ListCell<SoundDevice> treeCell, ListView<SoundDevice> treeView) {
+    private void drop(DragEvent event, ListCell<AudioDevice> treeCell, ListView<AudioDevice> treeView) {
         var db = event.getDragboard();
         var success = false;
         if (!db.hasContent(JAVA_FORMAT))
@@ -102,7 +102,7 @@ public class SoundDeviceImportFactory implements Callback<ListView<SoundDevice>,
             listView.getItems().add(index, draggedItem);
             treeView.getSelectionModel().select(draggedItem);
         } else {
-            listView.getItems().add(index, (SoundDevice) db.getContent(JAVA_FORMAT));
+            listView.getItems().add(index, (AudioDevice) db.getContent(JAVA_FORMAT));
         }
         event.setDropCompleted(success);
         clearDropLocation();
@@ -113,11 +113,11 @@ public class SoundDeviceImportFactory implements Callback<ListView<SoundDevice>,
             dropZone.setStyle("");
     }
 
-    private static void setupListView(ListView<SoundDevice> listView) {
+    private static void setupListView(ListView<AudioDevice> listView) {
         listView.setOnDragOver(event -> {
             if (!event.getDragboard().hasContent(JAVA_FORMAT) || !listView.getItems().isEmpty())
                 return;
-            var dropContent = (SoundDevice) event.getDragboard().getContent(JAVA_FORMAT);
+            var dropContent = (AudioDevice) event.getDragboard().getContent(JAVA_FORMAT);
             if (listView.getItems().contains(dropContent))
                 return;
             event.acceptTransferModes(TransferMode.MOVE);
@@ -128,7 +128,7 @@ public class SoundDeviceImportFactory implements Callback<ListView<SoundDevice>,
             if (!db.hasContent(JAVA_FORMAT) || !listView.getItems().isEmpty())
                 return;
             var index = listView.getItems().size();
-            listView.getItems().add(index, (SoundDevice) db.getContent(JAVA_FORMAT));
+            listView.getItems().add(index, (AudioDevice) db.getContent(JAVA_FORMAT));
             event.setDropCompleted(success);
         });
     }