+++ src/main/windows/sndctrl.h
@@ -1,4 +1,0 @@
-#pragma once
-
-void init();
-void destroy();